\name{getMACSPeaksAnno}
\alias{getMACSPeaksAnno}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Get MACS peaks annotated
}
\description{
This wrapper function takes the peak files generated by MACS2 as a input. Together with TxDb and Organism packages, it calls ChIPseeker::annotatePeak() to get peaks annotated.
}
\usage{
getMACSPeaksAnno(peakFile, genome = "mm",
                 txdb = NULL, annoDb = NULL,
		 rmsk = NULL, skip = 26,
		 annotate.repeat = TRUE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{peakFile}{A character string indicating the peak file genereated by MACS.
}
  \item{genome}{I should get rid of this argument...
}
  \item{txdb}{A TxDb object.}
  \item{annoDb}{A character string indicating the genome annotation package, such as org.hg.eg.db.}
  \item{rmsk}{A GRange instance of repeat masker (RMSK).
}
  \item{skip}{An integer specifying the head lines in the peak files subject to be skipped. Default to 26. Make sure to open the .xls peak file to confirm. If the skip lines is wrong, them the function would reture errors.
}
  \item{annotate.repeat}{A logical indicating whether to include repeats annoation.
}


\value{A GRanges instance with annotation.
}

\author{
Chao-Jen Wong
}

\examples{
\dontrun{
library(TxDb.Hsapiens.UCSC.hg19.knownGene)
mainDir <- "/fh/fast/tapscott_s/CompBio/ATAC-Seq/hg19.MB135.duxiB"
macsDir <- file.path(mainDir,"macs2")
peak_file <- list.files(macsDir, pattern=".xls$", full.names=TRUE)[1]
df <- getMACSPeaksAnno(peak_file
                       genome="hg", skip=21,
                       txdb=TxDb.Hsapiens.UCSC.hg19.knownGene,
                       annoDb="org.Hs.eg.db",
                       annotate.repeat=FALSE)
df
}
}


\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
