\name{SeqPipelineTools-package}
\alias{SeqPipelineTools-package}
\alias{SeqPipelineTools}
\docType{package}
\title{
\packageTitle{SeqPipelineTools}
}
\description{
\packageDescription{SeqPipelineTools}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SeqPipelineTools}
\packageIndices{SeqPipelineTools}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{SeqPipelineTools}

Maintainer: \packageMaintainer{SeqPipelineTools}
}

\keyword{ package }

